Folder rafctoring + MediatR implementation + model(Seniority) refactor + Added new records to Employee, Seniority, Roles
