+++ WarehouseManagementSystem/WarehouseManagementSystem.ApplicationServices/API/Domain/GetEmployeeRequest.cs
@@ -1,0 +1,13 @@
+﻿using MediatR;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace WarehouseManagementSystem.ApplicationServices.API.Domain
+{
+    public class GetEmployeeRequest : IRequest<GetEmployeeResponse>
+    {
+    }
+}
