+++ WarehouseManagementSystem/WarehouseManagementSystem.ApplicationServices/API/Domain/GetEmployeeResponse.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using WarehouseManagementSystem.ApplicationServices.API.Domain.Models;
+
+namespace WarehouseManagementSystem.ApplicationServices.API.Domain
+{
+    public class GetEmployeeResponse : ResponseBase<List<Employee>>
+    {
+    }
+}
