+++ WarehouseManagementSystem/WarehouseManagementSystem.ApplicationServices/API/Domain/Models/Employee.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace WarehouseManagementSystem.ApplicationServices.API.Domain.Models
+{
+    public class Employee
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public string LastName { get; set; }
+    }
+}
