+++ WarehouseManagementSystem/WarehouseManagementSystem.ApplicationServices/API/Domain/ResponseBase.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace WarehouseManagementSystem.ApplicationServices.API.Domain
+{
+    public class ResponseBase<T>
+    {
+        public T Data { get; set; }
+    }
+}
