+++ WarehouseManagementSystem/WarehouseManagementSystem/Controllers/EmployeeController.cs
@@ -1,4 +1,4 @@
-﻿using DataAccess;
+﻿using DataAccess.Repository;
 using DataAccess.Entities;
 using Microsoft.AspNetCore.Mvc;
 using System;
